pyproject.toml
[tool.poetry]
name = "nonebot-plugin-starrail-calendar"
version = "1.0.5"
description = "查看《崩坏：星穹铁道》官方活动"
authors = ["nicklly <1134741727@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_starrail_calendar"}]

[tool.poetry.dependencies]
python = "3.10.11"
nonebot2 = "2.0.0rc4"
nonebot-adapter-onebot = "2.2.2"
playwright = "1.32.1"
httpx = "0.23.3"
apscheduler = "3.10.1"
jinja2 = "3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
